+++ includes/MarketplaceWebService/Feed/SubmitFeed.php
@@ -21,7 +21,7 @@
   *
   */
 
-include_once ('.config.inc.php'); 
+include_once MWS_CONFIG;
 require_once WOO_AMZ_PLUGIN_DIR . 'includes/MarketplaceWebService/Client.php';
 
 $serviceUrl = "https://mws.amazonservices.com";