+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable theme changes should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [ 0.11.0 ] - Subscription Creation Stable, Subscription Update Stable, Test Notifications Stable - latest stable version
+
+## Added
+
+- Test Notification Support
+- Additional Model files for test notification support
+- Class MWS_Test_Notifications to handle testing notifications with registered subscription destinations
+- Update subscription support
+- Class MWS_Subscriptions to handle creating and updating Amazon MWS Subscriptions
+- Subscription Creation via Amazon SQS
+- Documentation for Subscription Creation
+- Amazon MWS Subscription service files and classes
+- Class MWS_Destination to handle registering and deregistered Amazon MWS Notification destinations
+
 ## [ 0.10.0 ]
 
 ## Notable features